requirements.txt
dspy>=2.5.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
RestrictedPython>=7.0
pydantic>=2.0.0
ragas>=0.2.0
sqlitedict>=2.1.0
aiohttp>=3.10.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
httpx>=0.27.0
pypdf>=4.0.0
python-docx>=1.0.0
langchain-text-splitters>=0.3.0
